requirements.txt
Flask==2.2.2
pymongo==4.4.1
python-dotenv==1.0.0
validators==0.20.0
